warnings: fix pedantic compiler warnings
Most of these are related to implicit sign conversion and width narrowing. They
have been "fixed" with casts in order to create a starting point for fixing
them correctly. Many of the changes in this commit are most definitely logically
_wrong_, but attempt to be consistent without breaking any functionality.

For example, the following old code:

  unsigned int a;
  int b = a;
  uint64_t c;
  uint32_t d = c;

Has become:

  unsigned int foo;
  int bar = (int)foo;
  uint64_t c;
  uint32_t d = (uint32_t)c;

This of course does nothing to fix any consequences of the narrowing and sign
conversions, but the casts become explicit.

By doing this as a first step, any attempt to fix the explicit casts more
correctly, while introducing a new warning, will be noticed by the pull-tester.

+++ src/ecmult_impl.h
@@ -109,7 +109,7 @@ static void secp256k1_ecmult_start(void) {
     secp256k1_ge_t ad = *g;
     secp256k1_gej_t fn; secp256k1_gej_set_infinity(&fn);
     for (int j=0; j<64; j++) {
-        for (int k=0; k<sizeof(secp256k1_ge_t); k++)
+        for (int k=0; k<(int)sizeof(secp256k1_ge_t); k++)
             ret->prec[j][k][0] = ((unsigned char*)(&ggn))[k];
         secp256k1_gej_add(&fn, &fn, &gg);
         for (int i=1; i<16; i++) {
@@ -117,7 +117,7 @@ static void secp256k1_ecmult_start(void) {
             secp256k1_ge_set_gej(&ggn, &gg);
             if (i == 15)
               ad = ggn;
-            for (int k=0; k<sizeof(secp256k1_ge_t); k++)
+            for (int k=0; k<(int)sizeof(secp256k1_ge_t); k++)
               ret->prec[j][k][i] = ((unsigned char*)(&ggn))[k];
         }
     }
@@ -183,7 +183,7 @@ void static secp256k1_ecmult_gen(secp256k1_gej_t *r, const secp256k1_num_t *gn) 
     for (int j=0; j<64; j++) {
         secp256k1_ge_t add;
         int bits = secp256k1_num_shift(&n, 4);
-        for (int k=0; k<sizeof(secp256k1_ge_t); k++)
+        for (int k=0; k<(int)sizeof(secp256k1_ge_t); k++)
             ((unsigned char*)(&add))[k] = c->prec[j][k][bits];
         secp256k1_gej_add_ge(r, r, &add);
     }
